refactor: move registerActionsCleanup
